refactor: changed file location to ui, and added base rounded style to fit project style-guide
