+++ src/features/register-player/game-styles/FormGameStyles.tsx
@@ -13,12 +13,12 @@ import { Button } from '@/components/ui/button'
 import { Arrow } from '@/components/icons/Arrow'
 
 import { cn } from '@/lib/utils'
+import { Skeleton } from '@/components/ui/Skeleton'
 import { validationCheckBoxLimitation } from './helpers/validationCheckBoxLimitation'
 import { FormAdditionalText } from './components/FormAdditionalText'
 import { FormTitle } from '../utils/title-form'
 import useSubmitGameStyles from './hooks/useSubmitGameStyles'
 import { useFormGameStyles } from './hooks/useFormGameStyles'
-import { Skeleton } from '../utils/Skeleton'
 import * as FormText from '../FormText.json'
 
 export function FormGameStyles() {
